Add condition for excluding ArgoCD Image Updater commits
ArgoCD Image Updater commits the new version of the image that is being used for updating the deployments. We need to prevent that commit from triggering a new image generation at Github Actions, otherwise we would be generating an infinite loop. What we do is to take the username of the commit for filtering those ones.